DRY up Team and Track Stream common code
Move common code into new class ActivityStream
